\name{datashield.assign}
\alias{datashield.assign}
\alias{datashield.assign.list}
\alias{datashield.assign.opal}
\title{Assign a value to a symbol in Opal(s).}
\usage{
  datashield.assign(object, ...)

  \method{datashield.assign}{opal} (opal, symbol, value)

  \method{datashield.assign}{list} (opals, ...)
}
\arguments{
  \item{opal}{}

  \item{symbol}{}

  \item{value}{}

  \item{opal}{opal object}

  \item{symbol}{}

  \item{value}{}

  \item{opals}{opal objects list}

  \item{symbol}{}

  \item{value}{}
}
\description{
  Assign a value to a symbol in Opal(s).

  Assign a value to a symbol in Opal.

  Assign a value to a symbol in each Opal.
}

