% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datashield.logout}
\alias{datashield.logout}
\title{Logout from Opal(s)}
\usage{
datashield.logout(opals)
}
\arguments{
\item{opals}{Opal object or a list of opals.}
}
\description{
Clear the Datashield R sessions and logout from Opal(s).
}

