% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.method.r
\name{datashield.method}
\alias{datashield.method}
\alias{datashield.method.list}
\alias{datashield.method.opal}
\title{Get Datashield method by its name}
\usage{
datashield.method(opal, name, type = "aggregate")

\method{datashield.method}{opal}(opal, name, type = "aggregate")

\method{datashield.method}{list}(opal, name, type = "aggregate")
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the method}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\description{
Get a Datashield method of a given type by its name.
}

