% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.r
\name{datashield.method_status}
\alias{datashield.method_status}
\title{Status of datashield method(s) in Opal(s)}
\usage{
datashield.method_status(opal, type = "aggregate")
}
\arguments{
\item{opal}{A list of opal objects.}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
Methods availability on each server.
}
\description{
Get the status of the datashield method(s) in different(s) Opal(s) server(s).
}

