% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.method.r
\name{datashield.methods}
\alias{datashield.methods}
\alias{datashield.methods.opal}
\alias{datashield.methods.list}
\title{List Datashield methods}
\usage{
datashield.methods(opal, type = "aggregate")

\method{datashield.methods}{opal}(opal, type = "aggregate")

\method{datashield.methods}{list}(opal, type = "aggregate")
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\description{
Get available Datashield methods of a given type.
}
