\name{datashield.newSession}
\alias{datashield.newSession}
\alias{datashield.newSession.list}
\alias{datashield.newSession.opal}
\title{Create a new Datashield session in Opal(s).}
\usage{
  datashield.newSession(opal)

  \method{datashield.newSession}{opal} (opal)

  \method{datashield.newSession}{list} (opals)
}
\arguments{
  \item{opal}{opal object or list of opal objects}

  \item{opal}{opal object}

  \item{opals}{opal objects list}
}
\description{
  Create a new Datashield session in Opal(s).

  Create a new Datashield session in Opal.

  Create a new Datashield session in each Opal.
}

