% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.r
\name{datashield.pkg_status}
\alias{datashield.pkg_status}
\title{Status of datashield package(s) in Opal(s)}
\usage{
datashield.pkg_status(opal)
}
\arguments{
\item{opal}{A list of opal objects.}
}
\value{
Packages status for each server.
}
\description{
Get the status of the datashield package(s) in different(s) Opal(s) server(s).
}
