% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.symbol.r
\name{datashield.rm}
\alias{datashield.rm}
\alias{datashield.rm.list}
\alias{datashield.rm.opal}
\title{Remove a R symbol}
\usage{
datashield.rm(opal, symbol)

\method{datashield.rm}{opal}(opal, symbol)

\method{datashield.rm}{list}(opal, symbol)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the current Datashield session.
}

