% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datashield.r
\name{datashield.rmSessions}
\alias{datashield.rmSessions}
\alias{datashield.rmSessions.list}
\alias{datashield.rmSessions.opal}
\title{Remove Datashield sessions}
\usage{
datashield.rmSessions(opal)

\method{datashield.rmSessions}{opal}(opal)

\method{datashield.rmSessions}{list}(opals)
}
\arguments{
\item{opals}{Opal object or list of opal objects.}
}
\description{
Remove Datashield sessions in Opal(s).
}

