\name{datashield.symbols}
\alias{datashield.symbols}
\alias{datashield.symbols.list}
\alias{datashield.symbols.opal}
\title{Get symbols from Opal(s).}
\usage{
  datashield.symbols(object, ...)

  \method{datashield.symbols}{opal} (opal)

  \method{datashield.symbols}{list} (opals)
}
\arguments{
  \item{opal}{}

  \item{opal}{}

  \item{opal}{}
}
\description{
  Get symbols from Opal(s).

  Get symbols from Opal.

  Get symbols from each Opal.
}

