% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.status.r
\name{datashield.table_status}
\alias{datashield.table_status}
\title{Status of table(s) in Opal(s)}
\usage{
datashield.table_status(opal, logins)
}
\arguments{
\item{opal}{A list of opal objects.}

\item{logins}{A dataframe table that holds login details.}
}
\value{
Servers and tables accessibility status
}
\description{
Get the status of the tables in differents Opals servers.
}

