% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{datashield.workspace_rm}
\alias{datashield.workspace_rm}
\alias{datashield.workspace_rm.list}
\alias{datashield.workspace_rm.opal}
\title{Remove a DataSHIELD workspace from a opal.}
\usage{
datashield.workspace_rm(opal, ...)

\method{datashield.workspace_rm}{opal}(opal, ws)

\method{datashield.workspace_rm}{list}(opals, ws)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{ws}{The workspace name}
}
\description{
Remove a DataSHIELD workspace from a opal.
}

