% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{datashield.workspace_save}
\alias{datashield.workspace_save}
\alias{datashield.workspace_save.list}
\alias{datashield.workspace_save.opal}
\title{Save current session in a DataSHIELD workspace.}
\usage{
datashield.workspace_save(opal, save)

\method{datashield.workspace_save}{opal}(opal, save)

\method{datashield.workspace_save}{list}(opal, save)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{save}{The workspace name}
}
\description{
Save current session in a DataSHIELD workspace.
}

