% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{datashield.workspaces}
\alias{datashield.workspaces}
\alias{datashield.workspaces.list}
\alias{datashield.workspaces.opal}
\title{Get the DataSHIELD workspaces.}
\usage{
datashield.workspaces(opal)

\method{datashield.workspaces}{opal}(opal)

\method{datashield.workspaces}{list}(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Get the DataSHIELD workspaces.
}

