% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.authorizationHeader}
\alias{.authorizationHeader}
\title{Constructs the value for the Authorization header}
\usage{
.authorizationHeader(username, password)
}
\description{
Constructs the value for the Authorization header
}
\keyword{internal}
