% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.delete}
\alias{.delete}
\title{Issues a DELETE request to opal for the specified resource}
\usage{
.delete(opal, ..., query = list())
}
\description{
Issues a DELETE request to opal for the specified resource
}
\keyword{internal}
