% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{.extractOpalCSRFToken}
\alias{.extractOpalCSRFToken}
\title{Extract CSRF token from cookie data frame.}
\usage{
.extractOpalCSRFToken(cookies)
}
\description{
Extract CSRF token from cookie data frame.
}
\keyword{internal}
