% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{.extractOpalSessionId}
\alias{.extractOpalSessionId}
\title{Extract opalsid from cookie data frame.}
\usage{
.extractOpalSessionId(cookies)
}
\description{
Extract opalsid from cookie data frame.
}
\keyword{internal}
