% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getPEMFilePath}
\alias{.getPEMFilePath}
\title{Extract absolute path to the pem file}
\usage{
.getPEMFilePath(pem, directory = "~/.ssh")
}
\description{
Extract absolute path to the pem file
}
\keyword{internal}
