% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{.getRSessionId}
\alias{.getRSessionId}
\title{Extract R session Id from opal object, create a new R session if not found.}
\usage{
.getRSessionId(opal)
}
\description{
Extract R session Id from opal object, create a new R session if not found.
}
\keyword{internal}
