% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.handleResponseOrCallback}
\alias{.handleResponseOrCallback}
\title{Process response with default handler or the provided one}
\usage{
.handleResponseOrCallback(opal, response, callback = NULL)
}
\description{
Process response with default handler or the provided one
}
\keyword{internal}
