% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.listToJson}
\alias{.listToJson}
\title{Simple transformation function of a list into a JSON object/array.}
\usage{
.listToJson(l)
}
\description{
Simple transformation function of a list into a JSON object/array.
}
\keyword{internal}
