% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.newProgress}
\alias{.newProgress}
\title{Create a new progress instance with default settings.}
\usage{
.newProgress(
  format = "  :what [:bar] :percent /:elapsed",
  clear = getOption("opal.progress.clear", FALSE),
  total,
  width = 100
)
}
\description{
Create a new progress instance with default settings.
}
\keyword{internal}
