% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{.newSession}
\alias{.newSession}
\title{Create a new R session in Opal.}
\usage{
.newSession(opal, restore = NULL, profile = NULL, wait = TRUE)
}
\description{
Create a new R session in Opal.
}
\keyword{internal}
