% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{.opal.login}
\alias{.opal.login}
\title{Create the opal object}
\usage{
.opal.login(
  username,
  password,
  token,
  url,
  opts = list(),
  profile = profile,
  restore = NULL
)
}
\description{
Create the opal object
}
\keyword{internal}
