% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.post}
\alias{.post}
\title{Post a request w/o body content}
\usage{
.post(opal, ..., query = list(), body = "",
  contentType = "application/x-rscript", callback = NULL)
}
\description{
Post a request w/o body content
}
\keyword{internal}
