% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.put}
\alias{.put}
\title{Put a request w/o body content}
\usage{
.put(opal, ..., query = list(), body = "",
  contentType = "application/x-rscript", callback = NULL)
}
\description{
Put a request w/o body content
}
\keyword{internal}
