% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.r
\name{.rmDatashieldSession}
\alias{.rmDatashieldSession}
\title{Remove a Datashield R session in Opal.}
\usage{
.rmDatashieldSession(opal, save = NULL)
}
\description{
Remove a Datashield R session in Opal.
}
\keyword{internal}
