% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.select}
\alias{.select}
\title{Returns a list r such that r[[i]] == l[[i]][field] for all i:length(l)}
\usage{
.select(l, field)
}
\description{
Returns a list r such that r[[i]] == l[[i]][field] for all i:length(l)
}
\keyword{internal}
