% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tickProgress}
\alias{.tickProgress}
\title{Output the progress status if option "datashield.progress" is allows to.}
\usage{
.tickProgress(progress, tokens = list())
}
\description{
Output the progress status if option "datashield.progress" is allows to.
}
\keyword{internal}
