% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.quota.R
\name{dsadmin.quota_exists}
\alias{dsadmin.quota_exists}
\title{Check a DataSHIELD usage quota exists}
\usage{
dsadmin.quota_exists(
  opal,
  metric = "EXECUTION_TIME",
  type = "system",
  subject = NULL
)
}
\arguments{
\item{opal}{Opal object.}

\item{metric}{The usage metric: 'EXECUTION_TIME' (default) or 'SESSION_TIME'.}

\item{type}{The type of subject: system (default), user or group.}

\item{subject}{The user name or the group name, not applicable when the type is system.}
}
\description{
Check whether a DataSHIELD usage quota was defined for a subject and a usage metric.
}
\details{
Requires DataSHIELD or system administration permissions.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
if (!dsadmin.quota_exists(o, type = 'group', subject = 'datashield'))
  dsadmin.quota_save(o, limit = 60, type = 'group', subject = 'datashield')
opal.logout(o)
}
}
\seealso{
Other DataSHIELD quotas:
\code{\link[=dsadmin.quota]{dsadmin.quota()}},
\code{\link[=dsadmin.quota_delete]{dsadmin.quota_delete()}},
\code{\link[=dsadmin.quota_enable]{dsadmin.quota_enable()}},
\code{\link[=dsadmin.quota_save]{dsadmin.quota_save()}},
\code{\link[=dsadmin.quota_usage]{dsadmin.quota_usage()}},
\code{\link[=dsadmin.quotas]{dsadmin.quotas()}},
\code{\link[=opal.quotas]{opal.quotas()}}
}
\concept{DataSHIELD quotas}
