% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmo.annotations.R
\name{harmo.annotations}
\alias{harmo.annotations}
\title{List the annotations}
\usage{
harmo.annotations(tibble, variables = NULL)
}
\arguments{
\item{tibble}{Tibble to be annotated}

\item{variables}{A character vector of variable names to be inspected. If NULL or empty, all
the columns of the tibble will be inspected.}
}
\value{
A data frame in long format (one row per annotation).
}
\description{
List the annotations of each of the variables.
}
