% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmo.dictionary.R
\name{harmo.update.dictionary}
\alias{harmo.update.dictionary}
\title{Update the dictionary of a Opal table}
\usage{
harmo.update.dictionary(opal, project, table, variables, categories = NULL)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table will be located.}

\item{table}{Destination table name.}

\item{variables}{A data frame with one row per variable (column name) and then one column per property/attribute (Opal Excel format).}

\item{categories}{A data frame with one row per category (columns variable and name) and then column per property/attribute (Opal Excel format). If there are
no categories, this parameter is optional.}
}
\description{
Directly update the dictionary of a Opal table with the provided dictionary.
}
