% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.installed_package}
\alias{oadmin.installed_package}
\title{Check package is installed}
\usage{
oadmin.installed_package(opal, pkg, profile = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}

\item{profile}{The R servers profile name to which operation applies. See also \link{opal.profiles}.}
}
\value{
TRUE if installed
}
\description{
Check package is installed
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.installed_package(o, 'xxx')
oadmin.installed_package(o, 'stats')
opal.logout(o)
}
}
\seealso{
Other administration functions:
\code{\link[=oadmin.install_bioconductor_package]{oadmin.install_bioconductor_package()}},
\code{\link[=oadmin.install_cran_package]{oadmin.install_cran_package()}},
\code{\link[=oadmin.install_devtools]{oadmin.install_devtools()}},
\code{\link[=oadmin.install_github_package]{oadmin.install_github_package()}},
\code{\link[=oadmin.install_local_package]{oadmin.install_local_package()}},
\code{\link[=oadmin.install_package]{oadmin.install_package()}},
\code{\link[=oadmin.installed_devtools]{oadmin.installed_devtools()}},
\code{\link[=oadmin.installed_packages]{oadmin.installed_packages()}},
\code{\link[=oadmin.package_description]{oadmin.package_description()}},
\code{\link[=oadmin.remove_package]{oadmin.remove_package()}}
}
\concept{administration functions}
