% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.log_sql}
\alias{oadmin.log_sql}
\title{Get Opal SQL API logs}
\usage{
oadmin.log_sql(opal, all = TRUE)
}
\arguments{
\item{opal}{Opal connection object.}

\item{all}{Get all or only latest log messages.}
}
\description{
Get Opal SQL API logs
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.log_sql(o)
opal.logout(o)
}
}
