% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.package_description}
\alias{oadmin.package_description}
\title{Get package description}
\usage{
oadmin.package_description(opal, pkg, fields = NULL, profile = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}

\item{fields}{A character vector giving the fields to extract from each package's DESCRIPTION file in addition to the default ones, or NULL (default). Unavailable fields result in NA values.}

\item{profile}{The R servers profile name to which operation applies. See also \link{opal.profiles}.}
}
\description{
Get package description
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.package_description(o, 'stats')
opal.logout(o)
}
}
\seealso{
Other administration functions:
\code{\link[=oadmin.install_bioconductor_package]{oadmin.install_bioconductor_package()}},
\code{\link[=oadmin.install_cran_package]{oadmin.install_cran_package()}},
\code{\link[=oadmin.install_devtools]{oadmin.install_devtools()}},
\code{\link[=oadmin.install_github_package]{oadmin.install_github_package()}},
\code{\link[=oadmin.install_local_package]{oadmin.install_local_package()}},
\code{\link[=oadmin.install_package]{oadmin.install_package()}},
\code{\link[=oadmin.installed_devtools]{oadmin.installed_devtools()}},
\code{\link[=oadmin.installed_package]{oadmin.installed_package()}},
\code{\link[=oadmin.installed_packages]{oadmin.installed_packages()}},
\code{\link[=oadmin.remove_package]{oadmin.remove_package()}}
}
\concept{administration functions}
