% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.perm}
\alias{oadmin.perm}
\title{Get the R permissions (deprecated)}
\usage{
oadmin.perm(opal)
}
\arguments{
\item{opal}{Opal connection object.}
}
\value{
A data.frame with columns: subject, type, permission
}
\description{
Deprecated, use \link{oadmin.r_perm}.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.r_perm_add(o, c('andrei', 'valentina'), 'user', 'use')
oadmin.r_perm(o)
oadmin.r_perm_delete(o, c('andrei', 'valentina'), 'user')
opal.logout(o)
}
}
