% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.annotations.R
\name{opal.annotate}
\alias{opal.annotate}
\title{Apply the annotations to a Opal table}
\usage{
opal.annotate(opal, datasource, table, annotations)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}

\item{annotations}{A data frame of annotations, with the expected columns: 'variable' (variable name),
'taxonomy' (the taxonomy name), 'vocabulary' (the vocabulary name) and 'term' (the term value, if NULL
of NA the annotation is removed).}
}
\description{
Set the provided annotations (as the one that can be retrieved from \link{opal.annotations})
to the table's data dictionary. Variables that do not exists in the table are ignored.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
annots <- opal.annotations(o, 'CPTP', 'Coreqx_final')
opal.annotate(o, 'CPTP', 'Cag_coreqx', annots)
opal.logout(o)
}
}
\seealso{
Other datasource functions:
\code{\link[=opal.annotations]{opal.annotations()}},
\code{\link[=opal.attribute_values]{opal.attribute_values()}},
\code{\link[=opal.datasource]{opal.datasource()}},
\code{\link[=opal.datasources]{opal.datasources()}},
\code{\link[=opal.table]{opal.table()}},
\code{\link[=opal.tables]{opal.tables()}},
\code{\link[=opal.valueset]{opal.valueset()}},
\code{\link[=opal.variable]{opal.variable()}},
\code{\link[=opal.variable_summary]{opal.variable_summary()}},
\code{\link[=opal.variables]{opal.variables()}}
}
\concept{datasource functions}
