% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.command.R
\name{opal.commands}
\alias{opal.commands}
\title{List the asynchronous commands}
\usage{
opal.commands(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the list of asynchronous R commands in the remote R session.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.commands(o)
opal.logout(o)
}
}
\seealso{
Other command functions:
\code{\link[=opal.command]{opal.command()}},
\code{\link[=opal.command_result]{opal.command_result()}},
\code{\link[=opal.command_rm]{opal.command_rm()}},
\code{\link[=opal.commands_rm]{opal.commands_rm()}}
}
\concept{command functions}
