% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.command.R
\name{opal.commands_rm}
\alias{opal.commands_rm}
\title{Remove all asynchronous commands}
\usage{
opal.commands_rm(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Remove all asynchronous R commands in the remote R session.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.commands_rm(o)
opal.logout(o)
}
}
\seealso{
Other command functions:
\code{\link[=opal.command]{opal.command()}},
\code{\link[=opal.command_result]{opal.command_result()}},
\code{\link[=opal.command_rm]{opal.command_rm()}},
\code{\link[=opal.commands]{opal.commands()}}
}
\concept{command functions}
