% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.R
\name{opal.datasource}
\alias{opal.datasource}
\title{Get a datasource}
\usage{
opal.datasource(opal, datasource)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}
}
\description{
Get a datasource
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.datasource(o, 'CNSIM')
opal.logout(o)
}
}
\seealso{
Other datasource functions:
\code{\link[=opal.annotate]{opal.annotate()}},
\code{\link[=opal.annotations]{opal.annotations()}},
\code{\link[=opal.attribute_values]{opal.attribute_values()}},
\code{\link[=opal.datasources]{opal.datasources()}},
\code{\link[=opal.table]{opal.table()}},
\code{\link[=opal.tables]{opal.tables()}},
\code{\link[=opal.valueset]{opal.valueset()}},
\code{\link[=opal.variable]{opal.variable()}},
\code{\link[=opal.variable_summary]{opal.variable_summary()}},
\code{\link[=opal.variables]{opal.variables()}}
}
\concept{datasource functions}
