% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{opal.delete}
\alias{opal.delete}
\title{Generic REST resource deletion.}
\usage{
opal.delete(opal, ..., query = list(), callback = NULL)
}
\arguments{
\item{opal}{Opal object.}

\item{...}{Resource path segments.}

\item{query}{Named list of query parameters.}

\item{callback}{A callback function to handle the response object.}
}
\description{
Generic REST resource deletion.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url = 'https://opal-demo.obiba.org')
opal.delete(o, 'some', 'resource')
opal.logout(o)
}
}
\seealso{
Other REST functions:
\code{\link[=opal.get]{opal.get()}},
\code{\link[=opal.post]{opal.post()}},
\code{\link[=opal.put]{opal.put()}}
}
\concept{REST functions}
