% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file.download}
\alias{opal.file.download}
\title{Download a file}
\usage{
opal.file.download(opal, source, destination = NULL)
}
\arguments{
\item{source}{Path to the file in the Opal file system.}

\item{destination}{Path to the file to be written. If ommitted, file with same name in the working directory will be written.}
}
\description{
Download a file or a folder from the Opal file system
}
\examples{
{
# download a file
opal.file.download(o, '/home/administrator/joins/join-src-3.csv')
opal.file.download(o, '/home/administrator/datashield/hop.R')
opal.file.download(o, '/home/administrator/datashield/CNSIM1.zip')

# download, create destination folder and rename file
opal.file.download(o, '/home/administrator/spss/DatabaseTest.sav', 'spss/test.sav')

# download a folder
opal.file.download(o, '/home/administrator/joins', 'joins.zip')
}
}

