% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file.read}
\alias{opal.file.read}
\title{Read a file}
\usage{
opal.file.read(opal, source, destination)
}
\arguments{
\item{source}{Path to the file in the R session workspace (must exists).}

\item{destination}{Path to the destination file or folder. Any required sub-folders will be created.}
}
\description{
Read a file from the R session workspace into the Opal file system
}
\examples{
{
# read into folder
opal.file.read(o,"DatabaseTest.sav", "/tmp")
# read and rename
opal.file.read(o,"test/DatabaseTest.sav", "/tmp/Test.sav")
# user home expansion
opal.file.read(o,"DatabaseTest.sav", "~/coucou/pwel.sav")
}
}

