% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.file.write}
\alias{opal.file.write}
\title{Write a file}
\usage{
opal.file.write(opal, source, destination = NULL)
}
\arguments{
\item{source}{Path to the file in the Opal file system (must exists and be accessible for the user).}

\item{destination}{Path to the destination file, relative to the R session workspace. Any required sub-folders will be created. If ommitted, file with same name will be written.}
}
\description{
Write a file from the Opal file system into the R session workspace
}

