\name{opal.function_status}
\alias{opal.function_status}
\title{Status of datashield function(s) in Opal(s)}
\usage{
  opal.function_status(opals, type = "aggregate")
}
\arguments{
  \item{TODO.}{}
}
\description{
  Get the status of the datashield function(s) in
  different(s) Opal(s) server(s).
}

