% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.getSessions}
\alias{opal.getSessions}
\title{Get R sessions}
\usage{
opal.getSessions(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\value{
The list of session identifiers.
}
\description{
Get all session identifiers in Opal.
}

