% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.newSession}
\alias{opal.newSession}
\title{New R session}
\usage{
opal.newSession(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\value{
The identifier of the session created.
}
\description{
Create a new R session in Opal.
}

