% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.R
\name{opal.profiles}
\alias{opal.profiles}
\title{List R profiles}
\usage{
opal.profiles(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\value{
The R profiles as a data.frame or a list
}
\description{
Each R profile corresponds one R servers cluster name. These profiles names can be provided
when login (see \link{opal.login}) and on some package administration operations.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url = 'https://opal-demo.obiba.org')
opal.profiles(o)
opal.logout(o)
}
}
\seealso{
Other connection functions:
\code{\link[=opal.login]{opal.login()}},
\code{\link[=opal.logout]{opal.logout()}}
}
\concept{connection functions}
