% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.project.R
\name{opal.project}
\alias{opal.project}
\title{Get a project}
\usage{
opal.project(opal, project)
}
\arguments{
\item{opal}{Opal object.}

\item{project}{Name of the project}
}
\description{
Get a project
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.project(o, 'datashield')
opal.logout(o)
}
}
\seealso{
Other project functions:
\code{\link[=opal.project_create]{opal.project_create()}},
\code{\link[=opal.project_delete]{opal.project_delete()}},
\code{\link[=opal.project_exists]{opal.project_exists()}},
\code{\link[=opal.projects]{opal.projects()}},
\code{\link[=opal.projects_databases]{opal.projects_databases()}}
}
\concept{project functions}
