% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.project.R
\name{opal.project_command}
\alias{opal.project_command}
\title{Get project task}
\usage{
opal.project_command(opal, project, id)
}
\arguments{
\item{opal}{Opal object.}

\item{project}{Name of the project.}

\item{id}{The project command ID.}
}
\value{
The command state object.
}
\description{
Get the project's task command object.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
id <- opal.project_backup(o, 'GREENSPACE', '/home/administrator/backup/GREENSPACE', wait = FALSE)
opal.project_command(opal, 'GREENSPACE', id)
opal.logout(o)
}
}
