% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.project.R
\name{opal.projects}
\alias{opal.projects}
\title{Get projects}
\usage{
opal.projects(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get projects
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.projects(o)
opal.logout(o)
}
}
\seealso{
Other project functions:
\code{\link[=opal.project]{opal.project()}},
\code{\link[=opal.project_create]{opal.project_create()}},
\code{\link[=opal.project_delete]{opal.project_delete()}},
\code{\link[=opal.project_exists]{opal.project_exists()}},
\code{\link[=opal.projects_databases]{opal.projects_databases()}}
}
\concept{project functions}
