% Generated by roxygen2 (4.0.1): do not edit by hand
\name{opal.report}
\alias{opal.report}
\title{Opal report}
\usage{
opal.report(input, output = NULL,
  boot_style = getOption("opal.report.style"), progress = FALSE,
  verbose = FALSE)
}
\arguments{
\item{input}{Path to the R markdown report file}

\item{ouput}{Directory path where to ouput the html report file. Default is the current working directory.}

\item{boot_style}{The Bootstrap style to use if character, if NULL uses the default markdown to HTML converter (no Bootstrap style), if TRUE a menu is shown with the available styles.}

\item{progress}{Knitr progress option}

\item{verbose}{Knitr verbose option}
}
\description{
Helper function for generating reports.
}

