% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.r
\name{opal.report_md}
\alias{opal.report_md}
\title{Turn a R markdown file to html.}
\usage{
opal.report_md(inputFile, boot_style, progress, verbose)
}
\description{
Turn a R markdown file to html.
}
\keyword{internal}

