% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.resource.R
\name{opal.resource_perm_add}
\alias{opal.resource_perm_add}
\title{Add or update a permission on a resource}
\usage{
opal.resource_perm_add(
  opal,
  project,
  resource,
  subject,
  type = "user",
  permission
)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{The project name.}

\item{resource}{The resource name.}

\item{subject}{A vector of subject identifiers: user names or group names (depending on the type).}

\item{type}{The type of subject: user (default) or group.}

\item{permission}{The permission to apply: view, administrate. The 'view' permission
is suitable for DataSHIELD operations.}
}
\description{
Add or update a permission on a resource
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.resource_perm_add(o, 'CNSIM', 'CNSIM1', c('andrei', 'valentina'), 'user', 'view')
opal.resource_perm(o, 'CNSIM', 'CNSIM1')
opal.resource_perm_delete(o, 'CNSIM', 'CNSIM1', c('andrei', 'valentina'), 'user')
opal.logout(o)
}
}
\seealso{
Other resource functions:
\code{\link[=opal.resource]{opal.resource()}},
\code{\link[=opal.resource_create]{opal.resource_create()}},
\code{\link[=opal.resource_delete]{opal.resource_delete()}},
\code{\link[=opal.resource_exists]{opal.resource_exists()}},
\code{\link[=opal.resource_extension_create]{opal.resource_extension_create()}},
\code{\link[=opal.resource_get]{opal.resource_get()}},
\code{\link[=opal.resource_perm]{opal.resource_perm()}},
\code{\link[=opal.resource_perm_delete]{opal.resource_perm_delete()}},
\code{\link[=opal.resources]{opal.resources()}},
\code{\link[=opal.resources_perm]{opal.resources_perm()}},
\code{\link[=opal.resources_perm_add]{opal.resources_perm_add()}},
\code{\link[=opal.resources_perm_delete]{opal.resources_perm_delete()}}
}
\concept{resource functions}
