\name{opal.rmSession}
\alias{opal.rmSession}
\title{Remove R session}
\usage{
  opal.rmSession(opal, sessionId = NULL)
}
\arguments{
  \item{opal}{Opal object.}

  \item{sessionId}{The identifier of the session to be
  removed. If omitted, current session is removed.}
}
\description{
  Remove R session from Opal.
}

