% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.session.R
\name{opal.session_get}
\alias{opal.session_get}
\title{Get the R session}
\usage{
opal.session_get(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\value{
The R session object.
}
\description{
Get the R session details if it exists
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.session(o, wait = FALSE)
opal.session_get(o)
opal.logout(o)
}
}
\seealso{
Other session functions:
\code{\link[=opal.session]{opal.session()}},
\code{\link[=opal.session_delete]{opal.session_delete()}},
\code{\link[=opal.session_events]{opal.session_events()}},
\code{\link[=opal.session_exists]{opal.session_exists()}},
\code{\link[=opal.session_running]{opal.session_running()}}
}
\concept{session functions}
