% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.session.R
\name{opal.session_ready}
\alias{opal.session_ready}
\title{Check if the R session is ready}
\usage{
opal.session_ready(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\value{
A logical value.
}
\description{
Check if the remote R session is ready to receive R commands. Fails if the session does not exist.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.session(o, wait = FALSE)
opal.session_ready(o)
opal.logout(o)
}
}
\seealso{
Other session functions: 
\code{\link{opal.session}()},
\code{\link{opal.session_delete}()},
\code{\link{opal.session_exists}()},
\code{\link{opal.session_get}()}
}
\concept{session functions}
