% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.setSession}
\alias{opal.setSession}
\title{Set R session}
\usage{
opal.setSession(opal, sessionId)
}
\arguments{
\item{opal}{Opal object.}

\item{sessionId}{The identifier of the session.}
}
\value{
The identifier of the session created.
}
\description{
Set current R session in Opal.
}

