% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.sql}
\alias{opal.sql}
\title{Execute a SQL query on the tables of a project}
\usage{
opal.sql(opal, project, query, id.name = "id", df = TRUE)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table will be located.}

\item{query}{The SQL query statement.}

\item{id.name}{The name of the column representing the entity identifiers. Default is 'id'.}

\item{df}{Returned value is a data.frame or a list. Default is TRUE.}
}
\value{
The lists of columns and rows, as a data.frame or a list (see df parameter).
}
\description{
The SQL query can apply to raw tables and/or views and require the permission to 
view the values of these tables.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.sql(o, 'CNSIM', 
  'select avg(LAB_HDL) as HDL_AVG, GENDER from CNSIM1 where LAB_HDL is not null group by GENDER')
opal.logout(o)
}
}
