% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.symbol.R
\name{opal.symbol_rm}
\alias{opal.symbol_rm}
\title{Remove a R symbol}
\usage{
opal.symbol_rm(opal, symbol)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the remote R session.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.symbol_rm(o, 'D')
opal.logout(o)
}
}
\seealso{
Other symbol functions:
\code{\link[=opal.rm]{opal.rm()}},
\code{\link[=opal.symbol_import]{opal.symbol_import()}},
\code{\link[=opal.symbol_save]{opal.symbol_save()}},
\code{\link[=opal.symbols]{opal.symbols()}}
}
\concept{symbol functions}
