% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.symbol.R
\name{opal.symbols}
\alias{opal.symbols}
\title{List R symbols}
\usage{
opal.symbols(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get the R symbols available in the remote R session.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.symbols(o)
opal.logout(o)
}
}
\seealso{
Other symbol functions:
\code{\link[=opal.rm]{opal.rm()}},
\code{\link[=opal.symbol_import]{opal.symbol_import()}},
\code{\link[=opal.symbol_rm]{opal.symbol_rm()}},
\code{\link[=opal.symbol_save]{opal.symbol_save()}}
}
\concept{symbol functions}
