% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.R
\name{opal.task_delete}
\alias{opal.task_delete}
\title{Delete a task}
\usage{
opal.task_delete(opal, id)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{Task identifier.}
}
\description{
Delete a specific task.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.task_delete(o, '1')
opal.logout(o)
}
}
\seealso{
Other task functions:
\code{\link[=opal.task]{opal.task()}},
\code{\link[=opal.task_cancel]{opal.task_cancel()}},
\code{\link[=opal.task_wait]{opal.task_wait()}},
\code{\link[=opal.tasks]{opal.tasks()}}
}
\concept{task functions}
