% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.R
\name{opal.taxonomies}
\alias{opal.taxonomies}
\title{Get taxonomies}
\usage{
opal.taxonomies(opal, locale = "en", df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{locale}{The language for labels (default is "en").}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get all taxonomies. A taxonomy describes the annotations that can be applied to the variables. 
Taxonomies also drive the variables search interface.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.taxonomies(o)
opal.logout(o)
}
}
\seealso{
Other taxonomy functions:
\code{\link[=opal.taxonomy]{opal.taxonomy()}},
\code{\link[=opal.taxonomy_delete]{opal.taxonomy_delete()}},
\code{\link[=opal.taxonomy_download]{opal.taxonomy_download()}},
\code{\link[=opal.taxonomy_upload]{opal.taxonomy_upload()}},
\code{\link[=opal.terms]{opal.terms()}},
\code{\link[=opal.vocabularies]{opal.vocabularies()}},
\code{\link[=opal.vocabulary]{opal.vocabulary()}}
}
\concept{taxonomy functions}
