% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.R
\name{opal.taxonomy_delete}
\alias{opal.taxonomy_delete}
\title{Delete a taxonomy}
\usage{
opal.taxonomy_delete(opal, taxonomy)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy.}
}
\description{
Delete a taxonomy, without failing if the taxonomy does not exist.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.taxonomy_delete(o, 'Mlstr_area')
opal.logout(o)
}
}
\seealso{
Other taxonomy functions:
\code{\link[=opal.taxonomies]{opal.taxonomies()}},
\code{\link[=opal.taxonomy]{opal.taxonomy()}},
\code{\link[=opal.taxonomy_download]{opal.taxonomy_download()}},
\code{\link[=opal.taxonomy_upload]{opal.taxonomy_upload()}},
\code{\link[=opal.terms]{opal.terms()}},
\code{\link[=opal.vocabularies]{opal.vocabularies()}},
\code{\link[=opal.vocabulary]{opal.vocabulary()}}
}
\concept{taxonomy functions}
