% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.term}
\alias{opal.term}
\title{Get a vocabulary term from a opal.}
\usage{
opal.term(opal, taxonomy, vocabulary, term)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{vocabulary}{Name of the vocabulary in the taxonomy}

\item{term}{Name of the term in the vocabulary}
}
\description{
Get a vocabulary term from a opal.
}
