% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.R
\name{opal.terms}
\alias{opal.terms}
\title{Get the terms of a vocabulary}
\usage{
opal.terms(opal, taxonomy, vocabulary, locale = "en", df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{vocabulary}{Name of the vocabulary in the taxonomy}

\item{locale}{The language for labels (default is "en").}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get all the terms of a vocabulary. The term describes the value of a variable annotation.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.terms(o, 'Mlstr_area', 'Lifestyle_behaviours')
opal.logout(o)
}
}
\seealso{
Other taxonomy functions:
\code{\link[=opal.taxonomies]{opal.taxonomies()}},
\code{\link[=opal.taxonomy]{opal.taxonomy()}},
\code{\link[=opal.taxonomy_delete]{opal.taxonomy_delete()}},
\code{\link[=opal.taxonomy_download]{opal.taxonomy_download()}},
\code{\link[=opal.taxonomy_upload]{opal.taxonomy_upload()}},
\code{\link[=opal.vocabularies]{opal.vocabularies()}},
\code{\link[=opal.vocabulary]{opal.vocabulary()}}
}
\concept{taxonomy functions}
