% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.execute.R
\name{opal.unload_package}
\alias{opal.unload_package}
\title{Unload package}
\usage{
opal.unload_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\description{
Unload package from the remote R session.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.unload_package(o, 'stats')
opal.logout(o)
}
}
\seealso{
Other execution functions:
\code{\link[=opal.execute]{opal.execute()}},
\code{\link[=opal.execute.source]{opal.execute.source()}},
\code{\link[=opal.load_package]{opal.load_package()}}
}
\concept{execution functions}
