% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.R
\name{opal.vocabularies}
\alias{opal.vocabularies}
\title{Get the vocabularies of a taxonomy}
\usage{
opal.vocabularies(opal, taxonomy, locale = "en", df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{locale}{The language for labels (default is "en").}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get all the vocabularies of a taxonomy. A vocabulary describes the possible values of variable annotations.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.vocabularies(o, 'Mlstr_area')
opal.logout(o)
}
}
\seealso{
Other taxonomy functions:
\code{\link[=opal.taxonomies]{opal.taxonomies()}},
\code{\link[=opal.taxonomy]{opal.taxonomy()}},
\code{\link[=opal.taxonomy_delete]{opal.taxonomy_delete()}},
\code{\link[=opal.taxonomy_download]{opal.taxonomy_download()}},
\code{\link[=opal.taxonomy_upload]{opal.taxonomy_upload()}},
\code{\link[=opal.terms]{opal.terms()}},
\code{\link[=opal.vocabulary]{opal.vocabulary()}}
}
\concept{taxonomy functions}
