% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.R
\name{opal.vocabulary}
\alias{opal.vocabulary}
\title{Get a taxonomy vocabulary}
\usage{
opal.vocabulary(opal, taxonomy, vocabulary)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{vocabulary}{Name of the vocabulary in the taxonomy}
}
\description{
Get a specific vocabulary details.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.vocabulary(o, 'Mlstr_area', 'Lifestyle_behaviours')
opal.logout(o)
}
}
\seealso{
Other taxonomy functions:
\code{\link[=opal.taxonomies]{opal.taxonomies()}},
\code{\link[=opal.taxonomy]{opal.taxonomy()}},
\code{\link[=opal.taxonomy_delete]{opal.taxonomy_delete()}},
\code{\link[=opal.taxonomy_download]{opal.taxonomy_download()}},
\code{\link[=opal.taxonomy_upload]{opal.taxonomy_upload()}},
\code{\link[=opal.terms]{opal.terms()}},
\code{\link[=opal.vocabularies]{opal.vocabularies()}}
}
\concept{taxonomy functions}
