% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.workspace.R
\name{opal.workspace_restore}
\alias{opal.workspace_restore}
\title{Restore a R workspace from a opal.}
\usage{
opal.workspace_restore(opal, ws)
}
\arguments{
\item{opal}{Opal object.}

\item{ws}{The workspace name}
}
\description{
Restore a R workspace from a opal.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.workspace_restore(o, 'test')
opal.logout(o)
}
}
\seealso{
Other workspace functions:
\code{\link[=opal.workspace_rm]{opal.workspace_rm()}},
\code{\link[=opal.workspace_save]{opal.workspace_save()}},
\code{\link[=opal.workspaces]{opal.workspaces()}}
}
\concept{workspace functions}
