% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.login.r
\name{datashield.logout}
\alias{datashield.logout}
\title{Logout from Opal(s)}
\usage{
datashield.logout(opals, save = NULL)
}
\arguments{
\item{opals}{Opal object or a list of opals.}

\item{save}{Save datashield sessions on each opal with provided ID (must be a character string).}
}
\description{
Clear the Datashield R sessions and logout from Opal(s).
}

