\name{datashield.method_status}
\alias{datashield.method_status}
\title{Status of datashield method(s) in Opal(s)}
\usage{
  datashield.method_status(opals, type = "aggregate")
}
\arguments{
  \item{TODO.}{}
}
\description{
  Get the status of the datashield method(s) in
  different(s) Opal(s) server(s).
}

