% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datashield.r
\name{datashield.newSession}
\alias{datashield.newSession}
\alias{datashield.newSession.list}
\alias{datashield.newSession.opal}
\title{New Datashield session}
\usage{
datashield.newSession(opals)

\method{datashield.newSession}{opal}(opal)

\method{datashield.newSession}{list}(opals)
}
\arguments{
\item{opals}{Opal object or list of opal objects.}
}
\value{
The identifier of the session created.
}
\description{
Create a new Datashield session in Opal(s).
}

