% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datashield.setSession}
\alias{datashield.setSession}
\alias{datashield.setSession.list}
\alias{datashield.setSession.opal}
\title{Set Datashield session}
\usage{
datashield.setSession(opal, sessionId)

\method{datashield.setSession}{opal}(opal, sessionId)

\method{datashield.setSession}{list}(opals, sessionId)
}
\arguments{
\item{sessionId}{The identifier of the session.}

\item{opals}{Opal object or list of opal objects.}
}
\description{
Set current Datashield session in Opal(s).
}

