\name{datashield.setSession}
\alias{datashield.setSession}
\alias{datashield.setSession.list}
\alias{datashield.setSession.opal}
\title{Set Datashield session}
\usage{
  datashield.setSession(opal, sessionId)

  \method{datashield.setSession}{opal} (opal, sessionId)

  \method{datashield.setession}{list} (opals, sessionId)
}
\arguments{
  \item{opals}{Opal object or list of opal objects.}

  \item{sessionId}{The identifier of the session.}
}
\description{
  Set current Datashield session in Opal(s).
}

