% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datashield.symbol.r
\name{datashield.symbols}
\alias{datashield.symbols}
\alias{datashield.symbols.list}
\alias{datashield.symbols.opal}
\title{List R symbols}
\usage{
datashield.symbols(opal, ...)

\method{datashield.symbols}{opal}(opal)

\method{datashield.symbols}{list}(opals)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Get the R symbols available after the datashield.assign calls in the current Datashield session.
}

