% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.get}
\alias{.get}
\title{Issues a GET request to opal for the specified resource}
\usage{
.get(opal, ..., query = list(), callback = NULL)
}
\description{
Issues a GET request to opal for the specified resource
}
\keyword{internal}
