% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{.handleResponseLocation}
\alias{.handleResponseLocation}
\title{Default request response Location handler.}
\usage{
.handleResponseLocation(opal, response)
}
\description{
Default request response Location handler.
}
\keyword{internal}
