% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.get_options}
\alias{dsadmin.get_options}
\title{Get Datashield Options}
\usage{
dsadmin.get_options(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Get the Datashield options.
}
