% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.install_package}
\alias{dsadmin.install_package}
\title{Install Datashield Package}
\usage{
dsadmin.install_package(opal, pkg, githubusername = NULL, ref = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}

\item{githubusername}{GitHub username of git repository. If NULL (default), try to install from Datashield package repository.}

\item{ref}{Desired git reference (could be a commit, tag, or branch name). If NULL (default), try to install from Datashield package repository.}
}
\value{
TRUE if installed
}
\description{
Install a package from Datashield public package repository or (if Git reference and GitHub username is provided) from Datashield source repository on GitHub.
}
