% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.installed_package}
\alias{dsadmin.installed_package}
\title{Check if a Datashield package is installed in Opal(s).}
\usage{
dsadmin.installed_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\value{
TRUE if installed
}
\description{
Check if a Datashield package is installed in Opal(s).
}
