% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.R
\name{dsadmin.profile_exists}
\alias{dsadmin.profile_exists}
\title{Check a DataSHIELD profile exists}
\usage{
dsadmin.profile_exists(opal, name)
}
\arguments{
\item{opal}{Opal object.}

\item{name}{Name of the profile.}
}
\description{
Check a DataSHIELD profile exists
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
if (!dsadmin.profile_exists(o, name = 'survival'))
  dsadmin.profile_create(o, name = 'survival', cluster = 'demo')
opal.logout(o)
}
}
\seealso{
Other DataSHIELD profiles: 
\code{\link{dsadmin.profile}()},
\code{\link{dsadmin.profile_access}()},
\code{\link{dsadmin.profile_create}()},
\code{\link{dsadmin.profile_delete}()},
\code{\link{dsadmin.profile_enable}()},
\code{\link{dsadmin.profile_init}()},
\code{\link{dsadmin.profile_perm}()},
\code{\link{dsadmin.profile_perm_add}()},
\code{\link{dsadmin.profile_perm_delete}()},
\code{\link{dsadmin.profile_rparser}()},
\code{\link{dsadmin.profiles}()}
}
\concept{DataSHIELD profiles}
