% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.R
\name{dsadmin.profile_init}
\alias{dsadmin.profile_init}
\title{Initialize a DataSHIELD profile}
\usage{
dsadmin.profile_init(opal, name)
}
\arguments{
\item{opal}{Opal object.}

\item{name}{Name of the profile.}
}
\description{
Clean the DataSHIELD's profile settings from all methods and options. This settings
can then be repopulated using \link{dsadmin.publish_package}, \link{dsadmin.set_package_methods} 
or \link{dsadmin.set_option}.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
dsadmin.profile_create(o, name = 'survival', cluster = 'demo')
dsadmin.profile_init(o, name = 'survival')
lapply(dsadmin.package_descriptions(o, profile = 'survival')$Package, function(p) {
  dsadmin.publish_package(o, p, profile = 'survival')
})
opal.logout(o)
}
}
\seealso{
Other DataSHIELD profiles: 
\code{\link{dsadmin.profile_access}()},
\code{\link{dsadmin.profile_create}()},
\code{\link{dsadmin.profile_delete}()},
\code{\link{dsadmin.profile_enable}()},
\code{\link{dsadmin.profile_exists}()},
\code{\link{dsadmin.profile_perm_add}()},
\code{\link{dsadmin.profile_perm_delete}()},
\code{\link{dsadmin.profile_perm}()},
\code{\link{dsadmin.profiles}()},
\code{\link{dsadmin.profile}()}
}
\concept{DataSHIELD profiles}
