% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.remove_package}
\alias{dsadmin.remove_package}
\title{Remove Datashield Package}
\usage{
dsadmin.remove_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\description{
Remove a Datashield package permanently from Opal(s).
}
