% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.rm_method}
\alias{dsadmin.rm_method}
\title{Remove Datashield Method}
\usage{
dsadmin.rm_method(opal, name, type = "aggregate")
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the method, as it is accessed by Datashield users.}

\item{type}{Type of the method: "aggregate" (default) or "assign"}
}
\description{
Remove a Datashield method from Opal(s).
}
