% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.rm_method}
\alias{dsadmin.rm_method}
\title{Remove DataSHIELD method}
\usage{
dsadmin.rm_method(opal, name, type = "aggregate")
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the method, as it is accessed by DataSHIELD users.}

\item{type}{Type of the method: "aggregate" (default) or "assign"}
}
\description{
Remove DataSHIELD method
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
dsadmin.rm_method(o, 'foo')
opal.logout(o)
}
}
\seealso{
Other DataSHIELD functions: 
\code{\link{dsadmin.get_methods}()},
\code{\link{dsadmin.get_method}()},
\code{\link{dsadmin.get_options}()},
\code{\link{dsadmin.install_package}()},
\code{\link{dsadmin.installed_package}()},
\code{\link{dsadmin.package_descriptions}()},
\code{\link{dsadmin.package_description}()},
\code{\link{dsadmin.remove_package}()},
\code{\link{dsadmin.rm_methods}()},
\code{\link{dsadmin.rm_option}()},
\code{\link{dsadmin.rm_package_methods}()},
\code{\link{dsadmin.set_method}()},
\code{\link{dsadmin.set_option}()},
\code{\link{dsadmin.set_package_methods}()}
}
\concept{DataSHIELD functions}
