% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.rm_methods}
\alias{dsadmin.rm_methods}
\title{Remove Datashield Methods}
\usage{
dsadmin.rm_methods(opal, type = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{type}{Type of the method: "aggregate" or "assign". Default is NULL (=all type of methods).}
}
\description{
Remove all Datashield methods from Opal(s).
}
