% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.rm_option}
\alias{dsadmin.rm_option}
\title{Remove Datashield Option}
\usage{
dsadmin.rm_option(opal, name)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the option}
}
\description{
Remove a Datashield option.
}
