% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.rm_package_methods}
\alias{dsadmin.rm_package_methods}
\title{Remove Package Datashield Methods}
\usage{
dsadmin.rm_package_methods(opal, pkg, type = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}

\item{type}{Type of the method: "aggregate" or "assign". Default is NULL (=all type of methods).}
}
\description{
Remove Datashield aggregate and assign methods defined by the package.
}
