% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.set_method}
\alias{dsadmin.set_method}
\title{Set Datashield Method}
\usage{
dsadmin.set_method(opal, name, func = NULL, path = NULL,
  type = "aggregate")
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the method, as it will be accessed by Datashield users.}

\item{func}{Function name.}

\item{path}{Path to the R file containing the script (mutually exclusive with func).}

\item{type}{Type of the method: "aggregate" (default) or "assign"}
}
\description{
Set a Datashield method in Opal(s).
}
