% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.set_option}
\alias{dsadmin.set_option}
\title{Set Datashield Option}
\usage{
dsadmin.set_option(opal, name, value)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the option}

\item{value}{Value of the option}
}
\description{
Set a Datashield option (add or update).
}
