% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.set_option}
\alias{dsadmin.set_option}
\title{Set DataSHIELD option}
\usage{
dsadmin.set_option(opal, name, value)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{name}{Name of the option}

\item{value}{Value of the option}
}
\description{
Set a DataSHIELD option (add or update).
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
dsadmin.set_option(o, 'foo', 'bar')
opal.logout(o)
}
}
\seealso{
Other DataSHIELD functions: \code{\link{dsadmin.get_methods}},
  \code{\link{dsadmin.get_method}},
  \code{\link{dsadmin.get_options}},
  \code{\link{dsadmin.install_package}},
  \code{\link{dsadmin.installed_package}},
  \code{\link{dsadmin.package_descriptions}},
  \code{\link{dsadmin.package_description}},
  \code{\link{dsadmin.remove_package}},
  \code{\link{dsadmin.rm_methods}},
  \code{\link{dsadmin.rm_method}},
  \code{\link{dsadmin.rm_option}},
  \code{\link{dsadmin.rm_package_methods}},
  \code{\link{dsadmin.set_method}},
  \code{\link{dsadmin.set_package_methods}}
}
\concept{DataSHIELD functions}
