% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.admin.r
\name{dsadmin.set_package_methods}
\alias{dsadmin.set_package_methods}
\title{Set Package Datashield Methods}
\usage{
dsadmin.set_package_methods(opal, pkg, type = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}

\item{type}{Type of the method: "aggregate" or "assign". Default is NULL (=all type of methods).}
}
\value{
TRUE if successfull
}
\description{
Declare Datashield aggregate and assign methods as defined by the package.
}
