% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.table.R
\name{harmo.table_get}
\alias{harmo.table_get}
\title{Get a Opal table as a tibble (deprecated)}
\usage{
harmo.table_get(opal, project, table, variables = NULL, missings = TRUE)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name from which the tibble should be extracted.}

\item{variables}{List of variable names or Javascript expression that selects the variables of a table (ignored if value does not refere to a table). See javascript documentation: http://wiki.obiba.org/display/OPALDOC/Variable+Methods}

\item{missings}{Include the missing values (default is TRUE).}
}
\description{
Deprecated: use \link{opal.table_get} instead.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
cqx <- opal.table_get(o, "CPTP", "Cag_coreqx")
opal.logout(o)
}
}
