% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.install_devtools}
\alias{oadmin.install_devtools}
\title{Install devtools package if not already available.}
\usage{
oadmin.install_devtools(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Install devtools package if not already available.
}
