% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.install_devtools}
\alias{oadmin.install_devtools}
\title{Install devtools package}
\usage{
oadmin.install_devtools(opal, profile = NULL)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{profile}{The R servers profile name to which operation applies. See also \link{opal.profiles}.}
}
\description{
Install devtools package if not already available.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.install_devtools(o)
opal.logout(o)
}
}
\seealso{
Other administration functions: 
\code{\link{oadmin.install_bioconductor_package}()},
\code{\link{oadmin.install_cran_package}()},
\code{\link{oadmin.install_github_package}()},
\code{\link{oadmin.install_local_package}()},
\code{\link{oadmin.install_package}()},
\code{\link{oadmin.installed_devtools}()},
\code{\link{oadmin.installed_package}()},
\code{\link{oadmin.installed_packages}()},
\code{\link{oadmin.package_description}()},
\code{\link{oadmin.remove_package}()}
}
\concept{administration functions}
