% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.installed_devtools}
\alias{oadmin.installed_devtools}
\title{Check if devtools package is installed.}
\usage{
oadmin.installed_devtools(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Check if devtools package is installed.
}
