% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.installed_devtools}
\alias{oadmin.installed_devtools}
\title{Check devtools package}
\usage{
oadmin.installed_devtools(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Check if devtools package is installed.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.installed_devtools(o)
opal.logout(o)
}
}
\seealso{
Other administration functions: 
\code{\link{oadmin.install_bioconductor_package}()},
\code{\link{oadmin.install_devtools}()},
\code{\link{oadmin.install_github_package}()},
\code{\link{oadmin.install_local_package}()},
\code{\link{oadmin.install_package}()},
\code{\link{oadmin.installed_packages}()},
\code{\link{oadmin.installed_package}()},
\code{\link{oadmin.package_description}()},
\code{\link{oadmin.remove_package}()}
}
\concept{administration functions}
