% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.installed_package}
\alias{oadmin.installed_package}
\title{Check if a package is installed in Opal(s).}
\usage{
oadmin.installed_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\value{
TRUE if installed
}
\description{
Check if a package is installed in Opal(s).
}
