% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.installed_package}
\alias{oadmin.installed_package}
\title{Check package is installed}
\usage{
oadmin.installed_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\value{
TRUE if installed
}
\description{
Check package is installed
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.installed_package(o, 'xxx')
oadmin.installed_package(o, 'stats')
opal.logout(o)
}
}
\seealso{
Other administration functions: 
\code{\link{oadmin.install_bioconductor_package}()},
\code{\link{oadmin.install_devtools}()},
\code{\link{oadmin.install_github_package}()},
\code{\link{oadmin.install_local_package}()},
\code{\link{oadmin.install_package}()},
\code{\link{oadmin.installed_devtools}()},
\code{\link{oadmin.installed_packages}()},
\code{\link{oadmin.package_description}()},
\code{\link{oadmin.remove_package}()}
}
\concept{administration functions}
