% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.installed_packages}
\alias{oadmin.installed_packages}
\title{Get the packages installed in Opal(s).}
\usage{
oadmin.installed_packages(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\value{
The result of the installed.packages() call
}
\description{
Get the packages installed in Opal(s).
}
