% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.installed_packages}
\alias{oadmin.installed_packages}
\title{List installed packages}
\usage{
oadmin.installed_packages(opal, profile = NULL)
}
\arguments{
\item{opal}{Opal object.}

\item{profile}{The R servers profile name to which operation applies. See also \link{opal.profiles}.}
}
\value{
The result of the installed.packages() call
}
\description{
Get the installed packages from all the R servers in the cluster described by the profile.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.installed_packages(o)
opal.logout(o)
}
}
\seealso{
Other administration functions: 
\code{\link{oadmin.install_bioconductor_package}()},
\code{\link{oadmin.install_cran_package}()},
\code{\link{oadmin.install_devtools}()},
\code{\link{oadmin.install_github_package}()},
\code{\link{oadmin.install_local_package}()},
\code{\link{oadmin.install_package}()},
\code{\link{oadmin.installed_devtools}()},
\code{\link{oadmin.installed_package}()},
\code{\link{oadmin.package_description}()},
\code{\link{oadmin.remove_package}()}
}
\concept{administration functions}
