% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.installed_packages}
\alias{oadmin.installed_packages}
\title{List installed packages}
\usage{
oadmin.installed_packages(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\value{
The result of the installed.packages() call
}
\description{
List installed packages
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
oadmin.installed_packages(o)
opal.logout(o)
}
}
\seealso{
Other administration functions: 
\code{\link{oadmin.install_devtools}()},
\code{\link{oadmin.install_github}()},
\code{\link{oadmin.install_package}()},
\code{\link{oadmin.installed_devtools}()},
\code{\link{oadmin.installed_package}()},
\code{\link{oadmin.package_description}()},
\code{\link{oadmin.remove_package}()}
}
\concept{administration functions}
