% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.perm}
\alias{oadmin.perm}
\title{Get the R permissions}
\usage{
oadmin.perm(opal)
}
\arguments{
\item{opal}{Opal connection object.}
}
\value{
A data.frame with columns: subject, type, permission
}
\description{
Get the permissions that were applied to the R service.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.perm_add(o, c('andrei', 'valentina'), 'user', 'use')
oadmin.perm(o)
oadmin.perm_delete(o, c('andrei', 'valentina'), 'user')
opal.logout(o)
}
}
