% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.R
\name{oadmin.perm_delete}
\alias{oadmin.perm_delete}
\title{Delete a R permission (deprecated)}
\usage{
oadmin.perm_delete(opal, subject, type = "user")
}
\arguments{
\item{opal}{Opal connection object.}

\item{subject}{A vector of subject identifiers: user names or group names (depending on the type).}

\item{type}{The type of subject: user (default) or group.}
}
\description{
Deprecated, use \link{oadmin.r_perm_delete}.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.r_perm_add(o, c('andrei', 'valentina'), 'user', 'use')
oadmin.r_perm(o)
oadmin.r_perm_delete(o, c('andrei', 'valentina'), 'user')
opal.logout(o)
}
}
