% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.remove_package}
\alias{oadmin.remove_package}
\title{Remove Package}
\usage{
oadmin.remove_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\description{
Remove package permanently from Opal(s).
}
