% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.admin.r
\name{oadmin.remove_package}
\alias{oadmin.remove_package}
\title{Remove package}
\usage{
oadmin.remove_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\description{
Remove package permanently.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
oadmin.remove_package(o, 'xxx')
opal.logout(o)
}
}
\seealso{
Other administration functions: 
\code{\link{oadmin.install_devtools}()},
\code{\link{oadmin.install_github}()},
\code{\link{oadmin.install_package}()},
\code{\link{oadmin.installed_devtools}()},
\code{\link{oadmin.installed_packages}()},
\code{\link{oadmin.installed_package}()},
\code{\link{oadmin.package_description}()}
}
\concept{administration functions}
