% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.users.R
\name{oadmin.users}
\alias{oadmin.users}
\title{Get the users}
\usage{
oadmin.users(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the users from the Opal internal users registry.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
oadmin.users(o)
opal.logout(o)
}
}
\seealso{
Other user functions: 
\code{\link{oadmin.user_add}()},
\code{\link{oadmin.user_delete}()},
\code{\link{oadmin.user_enable}()},
\code{\link{oadmin.user_exists}()},
\code{\link{oadmin.user_profile_delete}()},
\code{\link{oadmin.user_profiles}()},
\code{\link{oadmin.user_reset_password}()}
}
\concept{user functions}
