% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.assign.data}
\alias{opal.assign.data}
\title{Data assignment}
\usage{
opal.assign.data(opal, symbol, value, async = FALSE)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}

\item{value}{The R object to assign (data.frame, vector).}

\item{async}{R script is executed asynchronously within the session (default is FALSE). If TRUE, the value returned is the ID of the command to look for (from Opal 2.1).
#' @examples {
# push an arbitrary data frame to the R server
opal.assign.data(o, "D", mtcars)

# push an arbitrary vector to the R server
opal.assign.data(o, "C", mtcars$cyl)
}}
}
\description{
Assign a R object to a R symbol in the current R session.
}

