\name{opal.assign}
\alias{opal.assign}
\title{Data assignment}
\usage{
  opal.assign(opal, symbol, value)
}
\arguments{
  \item{opal}{Opal object.}

  \item{symbol}{Name of the R symbol.}

  \item{value}{Fully qualified name of a variable or a
  table in Opal.}
}
\description{
  Assign a Opal value to a R symbol in the current R
  session.
}

