% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.assign.table}
\alias{opal.assign.table}
\title{Data assignment}
\usage{
opal.assign.table(opal, symbol, value, variables = NULL, missings = FALSE,
  identifiers = NULL, withIds = FALSE, withTimestamps = FALSE,
  async = FALSE)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{symbol}{Name of the R symbol.}

\item{value}{The value to assign evaluated in the following order: a R expression, a function, a fully qualified name of a variable or a table in Opal or any other R object (data.frame, vector).}

\item{variables}{List of variable names or Javascript expression that selects the variables of a table (ignored if value does not refere to a table). See javascript documentation: http://wiki.obiba.org/display/OPALDOC/Variable+Methods}

\item{missings}{If TRUE, missing values will be pushed from Opal to R, default is FALSE. Ignored if value is an R expression.}

\item{identifiers}{Name of the identifiers mapping to use when assigning entities to R (from Opal 2.0).}

\item{withIds}{Add a vector representing the entity identifiers (from Opal 2.6). Default is FALSE.}

\item{withTimestamps}{Add a vector representing the creation and last update timestamps (from Opal 2.6). Default is FALSE.}

\item{async}{R script is executed asynchronously within the session (default is FALSE). If TRUE, the value returned is the ID of the command to look for (from Opal 2.1).}
}
\description{
Assign a Opal table to a R symbol in the current R session.
}
\examples{
{
# assign a list of variables from table HOP of opal object o
opal.assign.table(o, symbol="D", value="demo.HOP", variables=list("GENDER","LAB_GLUC"))

# assign a table HOP with a identifiers column
opal.assign.table(o, symbol="H", value="demo.HOP", withIds=TRUE)

# assign all the variables matching 'LAB' from table HOP of opal object o
opal.assign.table(o, symbol="D", value="demo.HOP", variables="name().matches('LAB_')")
}
}

