% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opal.command}
\alias{opal.command}
\title{Get an asynchronous command}
\usage{
opal.command(opal, id, wait = FALSE)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{R command ID.}

\item{wait}{Wait for the command to complete.}
}
\description{
Get an asynchronous R commands in the remote R session.
}

