% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.command.r
\name{opal.command}
\alias{opal.command}
\title{Get an asynchronous command}
\usage{
opal.command(opal, id, wait = FALSE)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{R command ID.}

\item{wait}{Wait for the command to complete.}
}
\description{
Get an asynchronous R commands in the remote R session.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.command(o, '1234')
opal.logout(o)
}
}
\seealso{
Other command functions: \code{\link{opal.command_result}},
  \code{\link{opal.command_rm}},
  \code{\link{opal.commands_rm}},
  \code{\link{opal.commands}}
}
\concept{command functions}
