% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.command.r
\name{opal.commands}
\alias{opal.commands}
\title{List the asynchronous commands}
\usage{
opal.commands(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get the list of asynchronous R commands in the remote R session.
}

