% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.command.r
\name{opal.commands_rm}
\alias{opal.commands_rm}
\title{Remove all asynchronous commands}
\usage{
opal.commands_rm(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Remove all asynchronous R commands in the remote R session.
}

