% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.command.r
\name{opal.commands_rm}
\alias{opal.commands_rm}
\title{Remove all asynchronous commands}
\usage{
opal.commands_rm(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Remove all asynchronous R commands in the remote R session.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.commands_rm(o)
opal.logout(o)
}
}
\seealso{
Other command functions: 
\code{\link{opal.command_result}()},
\code{\link{opal.command_rm}()},
\code{\link{opal.commands}()},
\code{\link{opal.command}()}
}
\concept{command functions}
