% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.datasource}
\alias{opal.datasource}
\title{Get a datasource from a opal.}
\usage{
opal.datasource(opal, datasource)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}
}
\description{
Get a datasource from a opal.
}
