\name{opal.datasource}
\alias{opal.datasource}
\title{Get a datasource from a opal.}
\usage{
  opal.datasource(opal, datasource, fields = NULL)
}
\arguments{
  \item{opal}{Opal object.}

  \item{datasource}{Name of the datasource.}

  \item{fields}{}
}
\description{
  Get a datasource from a opal.
}

