% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.datasource}
\alias{opal.datasource}
\title{Get a datasource}
\usage{
opal.datasource(opal, datasource)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}
}
\description{
Get a datasource
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.datasource(o, 'datashield')
opal.logout(o)
}
}
\seealso{
Other datasource functions: \code{\link{opal.attribute_values}},
  \code{\link{opal.datasources}},
  \code{\link{opal.projects}}, \code{\link{opal.project}},
  \code{\link{opal.tables}}, \code{\link{opal.table}},
  \code{\link{opal.valueset}},
  \code{\link{opal.variables}}, \code{\link{opal.variable}}
}
\concept{datasource functions}
