% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.datasources}
\alias{opal.datasources}
\title{Get datasources from a opal.}
\usage{
opal.datasources(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get datasources from a opal.
}

