% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.datasources}
\alias{opal.datasources}
\title{Get datasources}
\usage{
opal.datasources(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get datasources
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.datasources(o)
opal.logout(o)
}
}
\seealso{
Other datasource functions: \code{\link{opal.annotate}},
  \code{\link{opal.annotations}},
  \code{\link{opal.attribute_values}},
  \code{\link{opal.datasource}},
  \code{\link{opal.projects}}, \code{\link{opal.project}},
  \code{\link{opal.tables}}, \code{\link{opal.table}},
  \code{\link{opal.valueset}},
  \code{\link{opal.variables}}, \code{\link{opal.variable}}
}
\concept{datasource functions}
