\name{opal.execute}
\alias{opal.execute}
\title{Execute a R script.}
\usage{
  opal.execute(opal, script, session = TRUE)
}
\arguments{
  \item{opal}{Opal object.}

  \item{script}{R script to execute.}

  \item{session}{Execute in current R session (default is
  TRUE).}
}
\description{
  Execute a R script.
}

