% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opal.execute}
\alias{opal.execute}
\title{Execute a R script}
\usage{
opal.execute(opal, script, async = FALSE, session = TRUE)
}
\arguments{
\item{script}{R script to execute.}

\item{async}{R script is executed asynchronously within the session (default is FALSE). If TRUE, the value returned is the ID of the command to look for (from Opal 2.1).}

\item{session}{Execute in current R session (default is TRUE).}

\item{opals}{Opal object or list of opal objects.}
}
\description{
Execute a R script on Opal(s).
}

