% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.execute.r
\name{opal.execute}
\alias{opal.execute}
\title{Execute a R script}
\usage{
opal.execute(opal, script, async = FALSE, session = TRUE)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{script}{R script to execute.}

\item{async}{R script is executed asynchronously within the session (default is FALSE). If TRUE, the value returned is the ID of the command to look for (from Opal 2.1).}

\item{session}{Execute in current R session (default is TRUE).}
}
\description{
Execute a R script in the remote R session.
}
\seealso{
Other execution functions: \code{\link{opal.load_package}},
  \code{\link{opal.unload_package}}
}
\concept{execution functions}
