% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.file.download}
\alias{opal.file.download}
\title{Download a file}
\usage{
opal.file.download(opal, source, destination = NULL)
}
\arguments{
\item{source}{Path to the file in the Opal file system.}

\item{destination}{Path to the file to be written. If ommitted, file with same name in the working directory will be written.}
}
\description{
Download a file from the Opal file system
}

