% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.file}
\alias{opal.file}
\title{Get file content}
\usage{
opal.file(opal, path)
}
\arguments{
\item{path}{Path to the file in the Opal file system.}
}
\description{
Get file content from the Opal file system
}

