% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.file}
\alias{opal.file}
\title{Get a file}
\usage{
opal.file(opal, path)
}
\arguments{
\item{path}{Path to the file in the Opal file system.}
}
\description{
Get a file from its path in the Opal file system
}

