% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.file.read}
\alias{opal.file.read}
\title{Read a file}
\usage{
opal.file.read(opal, source, destination)
}
\arguments{
\item{source}{Path to the file in the R session workspace (must exists).}

\item{destination}{Path to the destination file or folder. Any required sub-folders will be created.}
}
\description{
Read a file from the R session workspace into the Opal file system
}

