% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.R
\name{opal.file_info}
\alias{opal.file_info}
\title{Get file information}
\usage{
opal.file_info(opal, path)
}
\arguments{
\item{opal}{Opal object.}

\item{path}{Path to the file in the Opal file system.}
}
\description{
Get details about a file from the Opal file system.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.file_info(o, '/home/administrator/joins/join-src-3.csv')
opal.logout(o)
}
}
\seealso{
Other file functions:
\code{\link[=opal.file]{opal.file()}},
\code{\link[=opal.file_cp]{opal.file_cp()}},
\code{\link[=opal.file_download]{opal.file_download()}},
\code{\link[=opal.file_ls]{opal.file_ls()}},
\code{\link[=opal.file_mkdir]{opal.file_mkdir()}},
\code{\link[=opal.file_mkdir_tmp]{opal.file_mkdir_tmp()}},
\code{\link[=opal.file_mv]{opal.file_mv()}},
\code{\link[=opal.file_read]{opal.file_read()}},
\code{\link[=opal.file_rm]{opal.file_rm()}},
\code{\link[=opal.file_unzip]{opal.file_unzip()}},
\code{\link[=opal.file_upload]{opal.file_upload()}},
\code{\link[=opal.file_write]{opal.file_write()}}
}
\concept{file functions}
