% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file_ls}
\alias{opal.file_ls}
\title{List content of a folder}
\usage{
opal.file_ls(opal, path)
}
\arguments{
\item{opal}{Opal object.}

\item{path}{Path to the folder in the Opal file system.}
}
\description{
List content of a folder in the Opal file system
}
\examples{
\dontrun{
# list content of a folder
opal.file_ls(o, '/home/userx/deliverables')
}
}
