% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file_ls}
\alias{opal.file_ls}
\title{List content of a folder}
\usage{
opal.file_ls(opal, path)
}
\arguments{
\item{opal}{Opal object.}

\item{path}{Path to the folder in the Opal file system.}
}
\description{
List content of a folder in the Opal file system.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
# list content of a folder
opal.file_ls(o, '/home/administrator')
opal.logout(o)
}
}
\seealso{
Other file functions: \code{\link{opal.file_cp}},
  \code{\link{opal.file_download}},
  \code{\link{opal.file_mkdir}},
  \code{\link{opal.file_mv}}, \code{\link{opal.file_read}},
  \code{\link{opal.file_rm}},
  \code{\link{opal.file_upload}},
  \code{\link{opal.file_write}}, \code{\link{opal.file}}
}
\concept{file functions}
