% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file_mkdir}
\alias{opal.file_mkdir}
\title{Make a folder}
\usage{
opal.file_mkdir(opal, path)
}
\arguments{
\item{opal}{Opal object.}

\item{path}{Path to the new folder in the Opal file system.}
}
\description{
Make a folder in the Opal file system
}
\examples{
\dontrun{
# make a folder
opal.file_mkdir(o, '/home/userx/deliverables')
}
}

