% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.R
\name{opal.file_mkdir_tmp}
\alias{opal.file_mkdir_tmp}
\title{Make a temporary folder}
\usage{
opal.file_mkdir_tmp(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\value{
The path of the created folder.
}
\description{
Make a user personal temporary folder in the Opal file system (make sure it does not exists).
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
# make a folder
path <- opal.file_mkdir_tmp(o)
opal.logout(o)
}
}
\seealso{
Other file functions:
\code{\link[=opal.file]{opal.file()}},
\code{\link[=opal.file_cp]{opal.file_cp()}},
\code{\link[=opal.file_download]{opal.file_download()}},
\code{\link[=opal.file_info]{opal.file_info()}},
\code{\link[=opal.file_ls]{opal.file_ls()}},
\code{\link[=opal.file_mkdir]{opal.file_mkdir()}},
\code{\link[=opal.file_mv]{opal.file_mv()}},
\code{\link[=opal.file_read]{opal.file_read()}},
\code{\link[=opal.file_rm]{opal.file_rm()}},
\code{\link[=opal.file_unzip]{opal.file_unzip()}},
\code{\link[=opal.file_upload]{opal.file_upload()}},
\code{\link[=opal.file_write]{opal.file_write()}}
}
\concept{file functions}
