% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file_read}
\alias{opal.file_read}
\title{Read a file}
\usage{
opal.file_read(opal, source, destination)
}
\arguments{
\item{opal}{Opal object.}

\item{source}{Path to the file in the R session workspace (must exists).}

\item{destination}{Path to the destination file or folder. Any required sub-folders will be created.}
}
\description{
Read a file from the R session workspace into the Opal file system
}
\examples{
\dontrun{
# read into folder
opal.file_read(o,"DatabaseTest.sav", "/tmp")
# read and rename
opal.file_read(o,"test/DatabaseTest.sav", "/tmp/Test.sav")
# user home expansion
opal.file_read(o,"DatabaseTest.sav", "~/coucou/pwel.sav")
}
}
