% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.file.r
\name{opal.file_upload}
\alias{opal.file_upload}
\title{Upload a file}
\usage{
opal.file_upload(opal, source, destination)
}
\arguments{
\item{opal}{Opal object.}

\item{source}{Path to the file in the local file system.}

\item{destination}{Path of the destination folder in the Opal file system.}
}
\description{
Upload a file into the Opal file system
}
