% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.get}
\alias{opal.get}
\title{Generic REST resource getter.}
\usage{
opal.get(opal, ..., query = list())
}
\arguments{
\item{opal}{Opal object.}

\item{...}{Resource path segments.}

\item{query}{Named list of query parameters.}
}
\description{
Generic REST resource getter.
}
