% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.execute.r
\name{opal.load_package}
\alias{opal.load_package}
\title{Load package}
\usage{
opal.load_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\description{
Load package in the remote R session.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.load_package(o, 'stats')
opal.logout(o)
}
}
\seealso{
Other execution functions: 
\code{\link{opal.execute.source}()},
\code{\link{opal.execute}()},
\code{\link{opal.unload_package}()}
}
\concept{execution functions}
