% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.logout}
\alias{opal.logout}
\title{Logout from Opal(s)}
\usage{
opal.logout(opals)
}
\arguments{
\item{opals}{Opal object or a list of opals.}
}
\description{
Clear the R sessions and logout from Opal(s).
}

