% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.logout}
\alias{opal.logout}
\title{Logout from Opal(s)}
\usage{
opal.logout(opals, save = FALSE)
}
\arguments{
\item{opals}{Opal object or a list of opals.}

\item{save}{Save the workspace with given (default is FALSE, current session ID if TRUE)}
}
\description{
Clear the R sessions and logout from Opal(s).
}

