% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.logout}
\alias{opal.logout}
\title{Logout from Opal(s)}
\usage{
opal.logout(opal, save = FALSE)
}
\arguments{
\item{opal}{Opal object or a list of opals.}

\item{save}{Save the workspace with given identifier (default value is FALSE, current session ID if TRUE).}
}
\description{
Clear the R sessions and logout from Opal(s).
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.logout(o)
}
}
\seealso{
Other connection functions: 
\code{\link{opal.login}()}
}
\concept{connection functions}
