% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.project}
\alias{opal.project}
\title{Get a project from a opal.}
\usage{
opal.project(opal, project)
}
\arguments{
\item{opal}{Opal object.}

\item{project}{Name of the project}
}
\description{
Get a project from a opal.
}
