% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.project}
\alias{opal.project}
\title{Get a project}
\usage{
opal.project(opal, project)
}
\arguments{
\item{opal}{Opal object.}

\item{project}{Name of the project}
}
\description{
Get a project
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.project(o, 'datashield')
opal.logout(o)
}
}
\seealso{
Other datasource functions: 
\code{\link{opal.annotate}()},
\code{\link{opal.annotations}()},
\code{\link{opal.attribute_values}()},
\code{\link{opal.datasources}()},
\code{\link{opal.datasource}()},
\code{\link{opal.projects}()},
\code{\link{opal.tables}()},
\code{\link{opal.table}()},
\code{\link{opal.valueset}()},
\code{\link{opal.variables}()},
\code{\link{opal.variable}()}
}
\concept{datasource functions}
