% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.projects}
\alias{opal.projects}
\title{Get projects from a opal.}
\usage{
opal.projects(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get projects from a opal.
}
