% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.projects}
\alias{opal.projects}
\title{Get projects}
\usage{
opal.projects(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get projects
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.projects(o)
opal.logout(o)
}
}
\seealso{
Other datasource functions: \code{\link{opal.annotate}},
  \code{\link{opal.annotations}},
  \code{\link{opal.attribute_values}},
  \code{\link{opal.datasources}},
  \code{\link{opal.datasource}},
  \code{\link{opal.project}}, \code{\link{opal.tables}},
  \code{\link{opal.table}}, \code{\link{opal.valueset}},
  \code{\link{opal.variables}}, \code{\link{opal.variable}}
}
\concept{datasource functions}
