% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{opal.report}
\alias{opal.report}
\title{Opal report}
\usage{
opal.report(input, output = NULL, progress = FALSE, verbose = FALSE,
  boot_style = NULL)
}
\arguments{
\item{input}{Path to the R markdown report file}

\item{progress}{Knitr progress option}

\item{verbose}{Knitr verbose option}

\item{boot_style}{Deprecated, directives can be integrated in the YAML header of the R markdown document.}

\item{ouput}{Directory path where to ouput the html report file. Default is the current working directory.}
}
\description{
Helper function for generating reports.
}

