% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.report.R
\name{opal.report}
\alias{opal.report}
\title{Opal report}
\usage{
opal.report(
  input,
  output = NULL,
  progress = FALSE,
  verbose = FALSE,
  boot_style = NULL
)
}
\arguments{
\item{input}{Path to the R markdown report file}

\item{output}{Directory path where to ouput the html report file. Default is the current working directory.}

\item{progress}{Knitr progress option}

\item{verbose}{Knitr verbose option}

\item{boot_style}{Deprecated, directives can be integrated in the YAML header of the R markdown document.}
}
\description{
Helper function for generating reports.
}
\examples{
\dontrun{
opal.report('input.Rmd', 'report', progress=TRUE)
}
}
