% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.resource.r
\name{opal.resource}
\alias{opal.resource}
\title{Get a resource of a project}
\usage{
opal.resource(opal, project, resource)
}
\arguments{
\item{opal}{Opal object.}

\item{project}{Name of the project.}

\item{resource}{Name of the resource in the project.}
}
\description{
Get a resource of a project
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.resource(o, 'datashield', 'CNSIM1r')
opal.logout(o)
}
}
\seealso{
Other project functions: 
\code{\link{opal.resources}()}
}
\concept{project functions}
