% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.resource.r
\name{opal.resources}
\alias{opal.resources}
\title{Get the resource references of a project}
\usage{
opal.resources(opal, project, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{project}{Name of the project.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the resource references of a project
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.resources(o, 'datashield')
opal.logout(o)
}
}
\seealso{
Other project functions: 
\code{\link{opal.resource}()}
}
\concept{project functions}
