% Generated by roxygen2 (4.0.1): do not edit by hand
\name{opal.rm}
\alias{opal.rm}
\title{Remove a R symbol (deprecated)}
\usage{
opal.rm(opal, symbol)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the current R session. Deprecated: see opal.symbol_rm function instead.
}

