% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.symbol.r
\name{opal.rm}
\alias{opal.rm}
\title{Remove a R symbol (deprecated)}
\usage{
opal.rm(opal, symbol)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the current R session. Deprecated: see opal.symbol_rm function instead.
}
