% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.rmSession}
\alias{opal.rmSession}
\title{Remove R session}
\usage{
opal.rmSession(opal, sessionId = NULL)
}
\arguments{
\item{opal}{Opal object.}

\item{sessionId}{The identifier of the session to be removed. If omitted, current session is removed.}
}
\description{
Remove R session from Opal.
}

