% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.session.R
\name{opal.session_exists}
\alias{opal.session_exists}
\title{Check if the R session exists}
\usage{
opal.session_exists(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\value{
A logical value.
}
\description{
Check if the remote R session exists
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.session_exists(o)
opal.logout(o)
}
}
\seealso{
Other session functions: 
\code{\link{opal.session}()},
\code{\link{opal.session_delete}()},
\code{\link{opal.session_events}()},
\code{\link{opal.session_get}()},
\code{\link{opal.session_running}()}
}
\concept{session functions}
