% Generated by roxygen2 (4.0.1): do not edit by hand
\name{opal.setSession}
\alias{opal.setSession}
\title{Set R session}
\usage{
opal.setSession(opal, sessionId)
}
\arguments{
\item{opal}{Opal object.}

\item{sessionId}{The identifier of the session.}
}
\value{
The identifier of the session created.
}
\description{
Set current R session in Opal.
}

