% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opal.symbol.r
\name{opal.symbol_rm}
\alias{opal.symbol_rm}
\title{Remove a R symbol}
\usage{
opal.symbol_rm(opal, symbol)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the current R session.
}

