% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.symbol.r
\name{opal.symbol_save}
\alias{opal.symbol_save}
\title{Data assignment}
\usage{
opal.symbol_save(opal, symbol, destination)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}

\item{destination}{The path of the file in the Opal file system. Supported file extensions are: .sav (SPSS), .sas7bdat (SAS), .dta (Stata).}
}
\description{
Save the tibble identified by the symbol in a file of the R session workspace
}

