% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.symbol.r
\name{opal.symbols}
\alias{opal.symbols}
\title{List R symbols}
\usage{
opal.symbols(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get the R symbols available in the remote R session.
}

