\name{opal.table}
\alias{opal.table}
\title{Get a table of a datasource from a opal.}
\usage{
opal.table(opal, datasource, table, fields = NULL)
}
\arguments{
  \item{opal}{Opal object.}

  \item{datasource}{Name of the datasource.}

  \item{table}{Name of the table in the datasource.}

  \item{fields}{}
}
\description{
Get a table of a datasource from a opal.
}

