% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.table}
\alias{opal.table}
\title{Get a table of a datasource from a opal.}
\usage{
opal.table(opal, datasource, table, counts = FALSE)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}

\item{counts}{Flag to get the number of variables and entities.}
}
\description{
Get a table of a datasource from a opal.
}

