% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.table.R
\name{opal.table_create}
\alias{opal.table_create}
\title{Create a Opal table}
\usage{
opal.table_create(opal, project, table, type = "Participant", silent = TRUE)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name to be deleted.}

\item{type}{Entity type, default is "Participant".}

\item{silent}{Warn if the table already exists, default is TRUE.}
}
\description{
Create a Opal table if it does not already exist.
}
\examples{
\dontrun{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.table_create(o, "CNSIM", "CNSIM4")
opal.logout(o)
}
}
