% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.table.R
\name{opal.table_dictionary_get}
\alias{opal.table_dictionary_get}
\title{Get the dictionary of a Opal table}
\usage{
opal.table_dictionary_get(opal, project, table)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table will be located.}

\item{table}{Table name.}
}
\description{
Get the dictionary of a Opal table in a format that can be re-applied with opal.table_dictionary_update.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
dico <- opal.table_dictionary_get(o, "CNSIM", "CNSIM1")
opal.logout(o)
}
}
