% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.table.R
\name{opal.table_exists}
\alias{opal.table_exists}
\title{Check a Opal table exists}
\usage{
opal.table_exists(opal, project, table)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name.}
}
\description{
Check whether a Opal table exists (and is visible).
}
\examples{
\dontrun{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.table_exists(o, "CNSIM", "CNSIM1")
opal.logout(o)
}
}
