% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.table.R
\name{opal.table_truncate}
\alias{opal.table_truncate}
\title{Truncate a Opal table}
\usage{
opal.table_truncate(opal, project, table)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name to be truncated.}
}
\description{
Removes the values of a table and keep the dictionary untouched. Fails if the table does
not exist or is a view. See also \link{opal.table_delete}.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.table_truncate(o, "CNSIM", "CNSIM1")
opal.logout(o)
}
}
