\name{opal.tables}
\alias{opal.tables}
\title{Get tables of a datasource from a opal.}
\usage{
opal.tables(opal, datasource, fields = NULL)
}
\arguments{
  \item{opal}{Opal object.}

  \item{datasource}{Name of the datasource.}

  \item{fields}{}
}
\description{
Get tables of a datasource from a opal.
}

