% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.r
\name{opal.tables}
\alias{opal.tables}
\title{Get tables of a datasource from a opal.}
\usage{
opal.tables(opal, datasource)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}
}
\description{
Get tables of a datasource from a opal.
}

