% Generated by roxygen2 (4.0.1): do not edit by hand
\name{opal.tables}
\alias{opal.tables}
\title{Get tables of a datasource from a opal.}
\usage{
opal.tables(opal, datasource)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}
}
\description{
Get tables of a datasource from a opal.
}

