% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.r
\name{opal.task}
\alias{opal.task}
\title{Get a task}
\usage{
opal.task(opal, id)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{Task identifier.}
}
\description{
Get the details of a specific task.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.task(o, '1')
opal.logout(o)
}
}
\seealso{
Other task functions: \code{\link{opal.task_cancel}},
  \code{\link{opal.task_wait}}, \code{\link{opal.tasks}}
}
\concept{task functions}
