% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.r
\name{opal.task_cancel}
\alias{opal.task_cancel}
\title{Tries to cancel a task from a opal.}
\usage{
opal.task_cancel(opal, id)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{Task identifier.}
}
\description{
Tries to cancel a task from a opal.
}
