% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.r
\name{opal.task_cancel}
\alias{opal.task_cancel}
\title{Cancel a task}
\usage{
opal.task_cancel(opal, id)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{Task identifier.}
}
\description{
Tries to cancel a task.
}
\seealso{
Other task functions: \code{\link{opal.task_wait}},
  \code{\link{opal.tasks}}, \code{\link{opal.task}}
}
\concept{task functions}
