% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.R
\name{opal.task_cancel}
\alias{opal.task_cancel}
\title{Cancel a task}
\usage{
opal.task_cancel(opal, id)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{Task identifier.}
}
\description{
Tries to cancel a task.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.task_cancel(o, '1')
opal.logout(o)
}
}
\seealso{
Other task functions: 
\code{\link{opal.task}()},
\code{\link{opal.task_wait}()},
\code{\link{opal.tasks}()}
}
\concept{task functions}
