% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.r
\name{opal.task_wait}
\alias{opal.task_wait}
\title{Wait for a task from a opal to complete.}
\usage{
opal.task_wait(opal, id, max = NULL)
}
\arguments{
\item{opal}{Opal object.}

\item{id}{Task identifier.}

\item{max}{Maximum time (in seconds) to wait for the task completion. Default is NULL (no maximum).}
}
\description{
Wait for a task from a opal to complete.
}
