% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.r
\name{opal.tasks}
\alias{opal.tasks}
\title{Get the tasks from a opal.}
\usage{
opal.tasks(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the tasks from a opal.
}
