% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.task.R
\name{opal.tasks}
\alias{opal.tasks}
\title{Get the tasks}
\usage{
opal.tasks(opal, df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get all the tasks with their status at the time of the request.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.tasks(o)
opal.logout(o)
}
}
\seealso{
Other task functions: 
\code{\link{opal.task_cancel}()},
\code{\link{opal.task_wait}()},
\code{\link{opal.task}()}
}
\concept{task functions}
