% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.taxonomies}
\alias{opal.taxonomies}
\title{Get taxonomies from a opal.}
\usage{
opal.taxonomies(opal, locale = "en", df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{locale}{The language for labels (default is "en").}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get taxonomies from a opal.
}
