% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.taxonomy}
\alias{opal.taxonomy}
\title{Get a taxonomy}
\usage{
opal.taxonomy(opal, taxonomy)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}
}
\description{
Get a specific taxonomy details.
}
\seealso{
Other taxonomy functions: \code{\link{opal.taxonomies}},
  \code{\link{opal.terms}}, \code{\link{opal.term}},
  \code{\link{opal.vocabularies}},
  \code{\link{opal.vocabulary}}
}
\concept{taxonomy functions}
