% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.taxonomy}
\alias{opal.taxonomy}
\title{Get a taxonomy from a opal.}
\usage{
opal.taxonomy(opal, taxonomy)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}
}
\description{
Get a taxonomy from a opal.
}
