% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.taxonomy}
\alias{opal.taxonomy}
\title{Get a taxonomy}
\usage{
opal.taxonomy(opal, taxonomy)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}
}
\description{
Get a specific taxonomy details.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.taxonomy(o, 'Mlstr_area')
opal.logout(o)
}
}
\seealso{
Other taxonomy functions: \code{\link{opal.taxonomies}},
  \code{\link{opal.terms}},
  \code{\link{opal.vocabularies}},
  \code{\link{opal.vocabulary}}
}
\concept{taxonomy functions}
