% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.token.R
\name{opal.token}
\alias{opal.token}
\title{Get a personal access token}
\usage{
opal.token(opal, name)
}
\arguments{
\item{opal}{Opal object.}

\item{name}{Name of the token}
}
\description{
Get a personal access token
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.token(o, 'sql-1')
opal.logout(o)
}
}
\seealso{
Other token functions: 
\code{\link{opal.token_datashield_create}()},
\code{\link{opal.token_delete}()},
\code{\link{opal.token_r_create}()},
\code{\link{opal.token_sql_create}()},
\code{\link{opal.tokens}()}
}
\concept{token functions}
