% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.token.R
\name{opal.token_delete}
\alias{opal.token_delete}
\title{Delete a personal access token}
\usage{
opal.token_delete(opal, name)
}
\arguments{
\item{opal}{Opal object.}

\item{name}{Name of the token}
}
\description{
Delete a personal access token permanently. Like for the other token functions, 
this operation requires the user to authenticate with username/password credentials.
}
\examples{
\dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.token_delete(o, 'sql-1')
opal.logout(o)
}
}
\seealso{
Other token functions: 
\code{\link{opal.token_datashield_create}()},
\code{\link{opal.token_r_create}()},
\code{\link{opal.token_renew}()},
\code{\link{opal.token_sql_create}()},
\code{\link{opal.tokens}()},
\code{\link{opal.token}()}
}
\concept{token functions}
