% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.execute.r
\name{opal.unload_package}
\alias{opal.unload_package}
\title{Unload package from the current session.}
\usage{
opal.unload_package(opal, pkg)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{pkg}{Package name.}
}
\description{
Unload package from the current session.
}
