% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.valueset}
\alias{opal.valueset}
\title{Get the values of an entity in a table.}
\usage{
opal.valueset(opal, datasource, table, identifier)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}

\item{identifier}{Entity identifier.}
}
\description{
Get the values of an entity in a table.
}
