% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.valueset}
\alias{opal.valueset}
\title{Get the values of an entity}
\usage{
opal.valueset(opal, datasource, table, identifier)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}

\item{identifier}{Entity identifier.}
}
\description{
Get the values of an entity in a table.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.valueset(o, 'datashield', 'CNSIM1', '1008573362')
opal.logout(o)
}
}
\seealso{
Other datasource functions: \code{\link{opal.attribute_values}},
  \code{\link{opal.datasources}},
  \code{\link{opal.datasource}},
  \code{\link{opal.projects}}, \code{\link{opal.project}},
  \code{\link{opal.tables}}, \code{\link{opal.table}},
  \code{\link{opal.variables}}, \code{\link{opal.variable}}
}
\concept{datasource functions}
