% Generated by roxygen2 (4.0.1): do not edit by hand
\name{opal.variable}
\alias{opal.variable}
\title{Get a variable of a table from a opal.}
\usage{
opal.variable(opal, datasource, table, variable)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}

\item{variable}{Name of the variable in the table.}
}
\description{
Get a variable of a table from a opal.
}

