% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.datasource.r
\name{opal.variables}
\alias{opal.variables}
\title{Get variables of a table from a opal.}
\usage{
opal.variables(opal, datasource, table, locale = "en", df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}

\item{locale}{The language for labels (default is "en").}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get variables of a table from a opal.
}
