\name{opal.variables}
\alias{opal.variables}
\title{Get variables of a table from a opal.}
\usage{
opal.variables(opal, datasource, table, fields = NULL)
}
\arguments{
  \item{opal}{Opal object.}

  \item{datasource}{Name of the datasource.}

  \item{table}{Name of the table in the datasource.}

  \item{fields}{}
}
\description{
Get variables of a table from a opal.
}

