% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opal.variables}
\alias{opal.variables}
\title{Get variables of a table from a opal.}
\usage{
opal.variables(opal, datasource, table)
}
\arguments{
\item{opal}{Opal object.}

\item{datasource}{Name of the datasource.}

\item{table}{Name of the table in the datasource.}
}
\description{
Get variables of a table from a opal.
}

