% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.vocabularies}
\alias{opal.vocabularies}
\title{Get the vocabularies of a taxonomy from a opal.}
\usage{
opal.vocabularies(opal, taxonomy, locale = "en", df = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{locale}{The language for labels (default is "en").}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the vocabularies of a taxonomy from a opal.
}
