% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.vocabulary}
\alias{opal.vocabulary}
\title{Get a taxonomy vocabulary}
\usage{
opal.vocabulary(opal, taxonomy, vocabulary)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{vocabulary}{Name of the vocabulary in the taxonomy}
}
\description{
Get a specific vocabulary details.
}
\seealso{
Other taxonomy functions: \code{\link{opal.taxonomies}},
  \code{\link{opal.taxonomy}}, \code{\link{opal.terms}},
  \code{\link{opal.term}}, \code{\link{opal.vocabularies}}
}
\concept{taxonomy functions}
