% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.taxonomy.r
\name{opal.vocabulary}
\alias{opal.vocabulary}
\title{Get a taxonomy vocabulary from a opal.}
\usage{
opal.vocabulary(opal, taxonomy, vocabulary)
}
\arguments{
\item{opal}{Opal object.}

\item{taxonomy}{Name of the taxonomy}

\item{vocabulary}{Name of the vocabulary in the taxonomy}
}
\description{
Get a taxonomy vocabulary from a opal.
}
