% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.workspace.r
\name{opal.workspace_rm}
\alias{opal.workspace_rm}
\title{Remove a R workspace from a opal.}
\usage{
opal.workspace_rm(opal, ws, user = NULL)
}
\arguments{
\item{opal}{Opal object.}

\item{ws}{The workspace name}

\item{user}{The user name associated to the worskpace. If not provided, the current user is applied.}
}
\description{
Remove a R workspace from a opal.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.workspace_rm(o, 'test')
opal.logout(o)
}
}
\seealso{
Other workspace functions: \code{\link{opal.workspace_save}},
  \code{\link{opal.workspaces}}
}
\concept{workspace functions}
