% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal.workspace.r
\name{opal.workspaces}
\alias{opal.workspaces}
\title{Get the R workspaces from a opal.}
\usage{
opal.workspaces(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get the R workspaces from a opal.
}
\examples{
\donttest{
o <- opal.login('administrator','password','https://opal-demo.obiba.org')
opal.workspaces(o)
opal.logout(o)
}
}
\seealso{
Other workspace functions: 
\code{\link{opal.workspace_rm}()},
\code{\link{opal.workspace_save}()}
}
\concept{workspace functions}
