% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opalr-package.R
\docType{package}
\name{opalr-package}
\alias{opalr}
\alias{opalr-package}
\title{opalr: 'Opal' Data Repository Client and 'DataSHIELD' Utils}
\description{
Data integration Web application for biobanks by 'OBiBa'. 'Opal' is the core database application for biobanks. Participant data, once collected from any data source, must be integrated and stored in a central data repository under a uniform model. 'Opal' is such a central repository. It can import, process, validate, query, analyze, report, and export data. 'Opal' is typically used in a research center to analyze the data acquired at assessment centres. Its ultimate purpose is to achieve seamless data-sharing among biobanks. This 'Opal' client allows to interact with 'Opal' web services and to perform operations on the R server side. 'DataSHIELD' administration tools are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/obiba/opalr/}
  \item \url{https://www.obiba.org/opalr/}
  \item \url{https://www.obiba.org/pages/products/opal/}
  \item \url{https://academic.oup.com/ije/article/46/5/1372/4102813}
  \item \url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1008880}
  \item \url{https://www.datashield.org/}
  \item Report bugs at \url{https://github.com/obiba/opalr/issues/}
}

}
\author{
\strong{Maintainer}: Yannick Marcon \email{yannick.marcon@obiba.org} (\href{https://orcid.org/0000-0003-0138-2023}{ORCID})

Other contributors:
\itemize{
  \item Amadou Gaye (\href{https://orcid.org/0000-0002-1180-2792}{ORCID}) [contributor]
  \item OBiBa group \email{info@obiba.org} [copyright holder]
}

}
\keyword{internal}
